Update track_transform to enable clip offset
When I downloaded the tools and tried using them, the keyframes would always be set at the wrong position in my VSE. When I read through the script I realized that it never accounts for the clip having an offset so I fixed it for my own purposes and am now sharing it here. I also deleted a variable which was never used in the script.